+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Capturing a diagnostics bundle will no longer fail if the extension fails to activate ([#2273](https://github.com/swiftlang/vscode-swift/pull/2273))
 - Stop assuming that `swift.path` will point at a directory named `bin` ([#2288](https://github.com/swiftlang/vscode-swift/pull/2288))
 - Fixed code coverage information only showing for the first test target when buildnig with swift-build ([#2296](https://github.com/swiftlang/vscode-swift/pull/2296))
+- Fixed test run never completing when stopping a debug session ([#2329](https://github.com/swiftlang/vscode-swift/pull/2329))
 
 ## 2.16.6 - 2026-06-04